test(repl): move TestPkChanged and add TestIsMinimalRowImage in utils_test.go
Three helpers (pkChanged, pkValueEqual, isMinimalRowImage) moved from
client.go to utils.go. Co-locate their tests:

- TestPkChanged moved from client_test.go to utils_test.go. Test body
  unchanged.
- TestIsMinimalRowImage added (none existed). Covers nil
  SkippedColumns, empty per-row slices (FULL image), a single
  non-empty per-row slice (MINIMAL/NOBLOB detected), and multiple
  skipped columns in one row.
- pkValueEqual is exercised indirectly through TestPkChanged's
  numeric-equivalence / byte-vs-string subtests, which match how the
  function is used in production. No direct test added.
